refactor(cli): split tui_handlers_inference.rs into inference/ submodules (#1144)
`koda-cli/src/tui_handlers_inference.rs` had grown to 1212 lines —
roughly double the 600-line guideline. Originally extracted from
`TuiContext::run_event_loop` in #447; the file kept absorbing new
concerns each epic-#1141 PR landed:

  - #1137 / PR #1142 added `drain_bounded` + 5 regression tests
  - #1138 / PR #1143 added the `SelectArm` enum + Draw arm + frame
    coalescing
  - #1200 / #1216 layered in the per-turn cancel cascade
  - #1210 added the bg-activity tap into the post-turn drain
  - #1211 added the slash-command guard + 5 tests
  - #1207 added the activity overlay routing

Each PR was reasonably scoped, but the cumulative result was a single
file mixing four entirely independent concerns: the inner select loop,
post-turn cleanup, terminal input handling, and engine→UI rendering.

## Split shape

```
koda-cli/src/inference/
├── mod.rs                  ( 31 lines) re-exports + module docs
├── select_loop.rs          (467 lines) SelectArm + drain_bounded +
│                                       run_inference_turn (the inner
│                                       tokio::select! loop) +
│                                       drain_tests
├── crossterm_handler.rs    (562 lines) handle_crossterm_event_inline +
│                                       handle_inference_key_inline +
│                                       is_slash_command_attempt +
│                                       truncate_preview +
│                                       slash_guard_tests
├── ui_handler.rs           (104 lines) handle_inference_ui_inline
└── post_turn.rs            (151 lines) post_turn_cleanup +
                                        maybe_auto_compact
```

Every file is under 600. The largest (`crossterm_handler.rs` at 562) is
on the radar but cohesive — it's the entire keystroke surface for a
single mode and splitting further would scatter the approval-vs-loop-cap-
vs-ask-user-vs-feedback-vs-general decision tree across files.

## Visibility & dispatch

  - `pub(crate) async fn run_inference_turn` keeps its signature; the
    one production caller in `tui_context::events_idle::dispatch_input`
    is unchanged.
  - The free helpers (`handle_crossterm_event_inline`,
    `handle_inference_ui_inline`) are `pub(super)` so `select_loop.rs`
    can dispatch into them without exposing them outside the
    `inference` module.
  - `drain_bounded`, `SelectArm`, `is_slash_command_attempt`,
    `truncate_preview` are private to their files. The two test
    modules (`drain_tests`, `slash_guard_tests`) follow the helpers
    they exercise.
  - Two `impl TuiContext` blocks (one in `select_loop.rs`, one in
    `post_turn.rs`) — same type, same crate, rustc stitches them.
    `post_turn_cleanup` is `pub(super)` so `select_loop.rs` can call
    it; `maybe_auto_compact` stays private (only `post_turn_cleanup`
    calls it).

## Behavior change

None. Pure refactor. Every line of executable code is byte-identical to
before; only the `mod` boundaries moved. Confirmed by `cargo test
--workspace` showing the same 2594 / 0 result as PR B (#1283).

## Doc updates

`queue_lanes.rs` had two stale references to `tui_handlers_inference.rs`
in module + function docs — updated to point at
`inference::crossterm_handler` instead.

## Validation

  cargo fmt --all -- --check                            ✓
  cargo clippy --workspace --all-targets -- -D warnings ✓
  RUSTDOCFLAGS=-Dwarnings cargo doc --workspace --no-deps ✓
  python3 scripts/check_tokio_test_flavor.py            ✓
  cargo test --workspace                                ✓ 2594 passed, 0 failed

Closes #1144.
Refs #1141 (TUI responsiveness epic).
