+++ koda-cli/src/lib.rs
@@ -31,6 +31,7 @@ pub(crate) mod headless;
 pub(crate) mod highlight;
 pub(crate) mod history_render;
 pub(crate) mod hyperlink;
+pub(crate) mod inference;
 pub(crate) mod input;
 pub(crate) mod md_render;
 pub(crate) mod mouse_select;
@@ -49,7 +50,6 @@ pub(crate) mod tui_app;
 pub(crate) mod tui_bg_tasks;
 pub(crate) mod tui_commands;
 pub(crate) mod tui_context;
-pub(crate) mod tui_handlers_inference;
 pub(crate) mod tui_mcp;
 pub(crate) mod tui_output;
 pub(crate) mod tui_render;
